project/backend/360_django/djangosite/requirements.txt
django==2.0.0
djangorestframework==3.9.2
djongo
sqlparse==0.2.4
markdown
django-filter
pygments # for highlighting
